Significant refactor of salter.
This change contains a lot of collective changes piled in with a major
refactor.. Sorry. =/

The most important change here is the ability for salt to work with
existing EC2 clusters rather than being limited to Classic EC2 only.

This also adds some functionality to make the config files easier to
read and understand since you can use "*" as a sgroup to automatically
expand out to the known groups.

This also includes a major refactor of the logging structure to try and
unify the approach to logs and what gets written. This is the first part
of a major overhaul that will be necessary.